Widen speed limit types from uint16 to uint32
The upload/download speed limits (s_maxupload, s_maxdownload) and slot
allocation (s_slotallocation) were stored as uint16, capping configurable
speed to 65534 kB/s (~524 Mbps) since 65535 is treated as unlimited.
This is inadequate for modern gigabit+ connections.

Changes:

Preferences.h / Preferences.cpp:
- Change s_maxupload, s_maxdownload, s_slotallocation from uint16 to
  uint32, along with their getters and setters
- The existing backwards compatibility check in CheckUlDlRatio() that
  maps 0xFFFF to UNLIMITED still works correctly with uint32

UploadQueue.h / UploadQueue.cpp:
- Change GetMaxSlots() return type and local variables from uint16 to
  uint32 to prevent truncation when calculating slot counts from large
  speed values

RemoteConnect.h:
- Update SetPreferencesConnections() parameter types for
  MaxDownloadSpeed, MaxUploadSpeed, and UploadSlotAllocation from
  uint16 to uint32

EC protocol note:
- EC tags EC_TAG_CONN_MAX_UL, EC_TAG_CONN_MAX_DL, and
  EC_TAG_CONN_SLOT_ALLOCATION now carry uint32 values. EC clients
  reading these tags should use GetInt() which handles any integer
  size. Clients sending these tags should encode them as 32-bit.

+++ src/UploadQueue.h
@@ -69,7 +69,7 @@ public:
 
 private:
 	void	RemoveFromWaitingQueue(CClientRefList::iterator pos);
-	uint16	GetMaxSlots() const;
+	uint32	GetMaxSlots() const;
 	void	AddUpNextClient(CUpDownClient* directadd = 0);
 	bool	IsSuspended(const CMD4Hash& hash) { return suspendedUploadsSet.find(hash) != suspendedUploadsSet.end(); }
 	void	SortGetBestClient(CClientRef * bestClient = NULL);
